+++ lib/amazonka/src/Amazonka/EC2/Metadata.hs
@@ -60,6 +60,7 @@ import qualified Data.ByteString.Char8 as BS8
 import qualified Data.ByteString.Lazy as LBS
 import qualified Data.Text as Text
 import qualified Network.HTTP.Client as Client
+import Network.HTTP.Simple (setRequestHeader, setRequestMethod)
 
 data Dynamic
   = -- | Value showing whether the customer has enabled detailed one-minute
@@ -295,7 +296,7 @@ isEC2 :: MonadIO m => Client.Manager -> m Bool
 isEC2 m = liftIO (Exception.catch req err)
   where
     req = do
-      !_ <- request m "http://instance-data/latest"
+      !_ <- get m "http://instance-data/latest"
 
       return True
 
@@ -444,15 +445,31 @@ identity ::
 identity m = eitherDecode . LBS.fromStrict <$> dynamic m Document
 
 get :: MonadIO m => Client.Manager -> Text -> m ByteString
-get m url = liftIO (strip <$> request m url)
+get m url = liftIO $ do
+  token <- strip <$> requestToken
+  strip <$> requestWith (addToken token) m url
   where
+    requestToken =
+      requestWith
+        ( setRequestMethod "PUT"
+            . setRequestHeader "X-aws-ec2-metadata-token-ttl-seconds" ["60"]
+        )
+        m
+        (latest <> "api/token")
+
+    addToken token = setRequestHeader "X-aws-ec2-metadata-token" [token]
+
     strip bs
       | BS8.isSuffixOf "\n" bs = BS8.init bs
       | otherwise = bs
 
-request :: Client.Manager -> Text -> IO ByteString
-request m url = do
+requestWith ::
+  (Client.Request -> Client.Request) ->
+  Client.Manager ->
+  Text ->
+  IO ByteString
+requestWith modifyRequest m url = do
   rq <- Client.parseUrlThrow (Text.unpack url)
-  rs <- Client.httpLbs rq m
+  rs <- Client.httpLbs (modifyRequest rq) m
 
   return . LBS.toStrict $ Client.responseBody rs